refactor(sign-up): move widgets in SignUpForm to   SignUpViewBody
